move PrecisionChart to release package
